requirements.txt
numpy
torch
fastapi
kafka-python